requirements.txt
python-telegram-bot[webhooks]>=21.6
openai>=1.52.0
httpx>=0.27.2
